backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
slowapi==0.1.9

# AI & External APIs
openai==1.10.0
fastapi-mail==1.4.1

# Performance & Scalability
asyncp g==0.31.0
redis==7.1.0
celery==5.5.3
gunicorn==23.0.0
